passwd: Move password getting code to own file for sharing.
