+++ apps/mosquitto_passwd/CMakeLists.txt
@@ -5,6 +5,7 @@ include_directories(${mosquitto_SOURCE_DIR} ${mosquitto_SOURCE_DIR}/include
 if (WITH_TLS)
 	add_executable(mosquitto_passwd
 		mosquitto_passwd.c
+		get_password.c get_password.h
 		../../lib/memory_mosq.c ../../lib/memory_mosq.h
 		../../src/memory_public.c
 		../../lib/misc_mosq.c
